resolved issue #9 - refactor helper methods out of builderlayout
