+++ src/pages/AddCommunityPage/BuilderLayout/actions/index.js
@@ -1,1 +1,5 @@
 export * from './addElementMethod';
+export * from './moveElement';
+export * from './sharedMethods';
+export * from './updateActiveElement';
+export * from './updateActiveElementAttributes';
